#974 Eye Form v3
https://sourceforge.net/p/oscarmcmaster/feature-requests/974/

+++ src/main/java/org/oscarehr/eyeform/web/PlanAction.java
@@ -120,11 +120,11 @@ public class PlanAction extends DispatchAction {
     			//if(timespan.length() == 0) {
     			//	continue;
     			//}
-    			try {
-    				Integer.parseInt(timespan);
-    			}catch(NumberFormatException e) {
-    				timespan="0";
-    			}
+//    			try {
+//    				Integer.parseInt(timespan);
+//    			}catch(NumberFormatException e) {
+//    				timespan="0";
+//    			}
     			EyeformFollowUp fu = new EyeformFollowUp();
     			if(id.length()>0 && Integer.parseInt(id)>0) {
     				fu = followUpDao.find(Integer.parseInt(id));
@@ -136,7 +136,8 @@ public class PlanAction extends DispatchAction {
     			fu.setComment(request.getParameter("followup_"+x+".comment"));
     			fu.setFollowupProvider(request.getParameter("followup_"+x+".followupProvider"));
     			fu.setTimeframe(request.getParameter("followup_"+x+".timeframe"));
-    			fu.setTimespan(Integer.parseInt(timespan));
+//    			fu.setTimespan(Integer.parseInt(timespan));
+    			fu.setTimespan(timespan);
     			fu.setType(request.getParameter("followup_"+x+".type"));
     			fu.setUrgency(request.getParameter("followup_"+x+".urgency"));
 
@@ -151,8 +152,10 @@ public class PlanAction extends DispatchAction {
     	String[] ids = request.getParameterValues("followup.delete");
     	if(ids != null) {
     		for(String id:ids) {
-    			int followUpId = Integer.parseInt(id);
-    			followUpDao.remove(followUpId);
+    			if(id.length()>0) {
+        			int followUpId = Integer.parseInt(id);
+        			followUpDao.remove(followUpId);    				
+    			}
     		}
     	}
 
@@ -188,8 +191,10 @@ public class PlanAction extends DispatchAction {
     	ids = request.getParameterValues("procedure.delete");
     	if(ids != null) {
     		for(String id:ids) {
-    			int procedureId = Integer.parseInt(id);
-    			procBookDao.remove(procedureId);
+    			if(id.length()>0) {
+        			int procedureId = Integer.parseInt(id);
+        			procBookDao.remove(procedureId);    				
+    			}
     		}
     	}
 
@@ -225,8 +230,10 @@ public class PlanAction extends DispatchAction {
     	ids = request.getParameterValues("test.delete");
     	if(ids != null) {
     		for(String id:ids) {
-    			int testId = Integer.parseInt(id);
-    			testBookDao.remove(testId);
+    			if(id.length()>0) {
+    				int testId = Integer.parseInt(id);
+    				testBookDao.remove(testId);
+    			}
     		}
     	}
 
@@ -241,12 +248,14 @@ public class PlanAction extends DispatchAction {
 	    	for(String id:ids) {
 	    		EyeformFollowUp followUp = followUpDao.find(Integer.parseInt(id));
 	    		followUp.setType(request.getParameter("followup"+id+".type"));
-	    		followUp.setTimespan(Integer.parseInt(request.getParameter("followup"+id+".timespan")));
+//	    		followUp.setTimespan(Integer.parseInt(request.getParameter("followup"+id+".timespan")));
+	    		followUp.setTimespan(request.getParameter("followup"+id+".timespan"));
 	    		followUp.setFollowupProvider(request.getParameter("followup"+id+".followupProvider"));
 	    		followUp.setUrgency(request.getParameter("followup"+id+".urgency"));
 	    		followUp.setComment(request.getParameter("followup"+id+".comment"));
 	    		followUp.setTimeframe(request.getParameter("followup"+id+".timeframe"));
-	    		if(followUp.getTimespan()==0) {
+//	    		if(followUp.getTimespan()==0) {
+	    		if(followUp.getTimespan().equals("0") || followUp.getTimespan().equals("")) {
 	    			followUpDao.remove(Integer.parseInt(id));
 	    		} else {
 	    			followUpDao.merge(followUp);
